requirements.txt
shiny
ibis-framework[duckdb]
palmerpenguins
pymongo
pyairtable
python-dotenv
chatlas
querychat
